pip-requirements.txt
python-telegram-bot==13.6
Pillow==8.3.0